refactor: update import paths to use @sidequest/core
